Move under internalimageutil under internal/imageutil
