+++ pkg/image/apiserver/internal/imageutil/helpers.go
@@ -1,4 +1,4 @@
-package internalimageutil
+package imageutil
 
 import (
 	"encoding/json"
